Move helpers to GitHubOrg
